Implement career-ops-tracker Dashboard.vue and fix score column type
- Real Dashboard.vue: 4 stat cards (applications, funnel, pipeline
  health, scan activity) bound to career-ops's actual stats.mjs --json
  shape, plus a read-only applications table with status badges
- New shadcn-vue primitives: Badge (status colors) and Table
- Bug fix: score is a float (e.g. 3.5/5 from applications.md), not an
  integer as originally migrated — corrected column type, model cast,
  and validation rule; verified end-to-end via manual sync + browser
- Removed scaffold leftovers no longer applicable to this single-user
  read-only tool: register page/route (no self-registration), starter
  kit branding and Github/Docs footer links, unused placeholder art

+++ my-app/app/Models/Application.php
@@ -23,7 +23,7 @@ final class Application extends Model
 
     protected $casts = [
         'num' => 'integer',
-        'score' => 'integer',
+        'score' => 'float',
         'date' => 'date',
     ];
